+++ tests/TestCase.php
@@ -2,36 +2,32 @@
 
 namespace Tim661811\LaravelTelemetryReporter\Tests;
 
-use Illuminate\Database\Eloquent\Factories\Factory;
 use Orchestra\Testbench\TestCase as Orchestra;
 use Tim661811\LaravelTelemetryReporter\LaravelTelemetryReporterServiceProvider;
 
 class TestCase extends Orchestra
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'Tim661811\\LaravelTelemetryReporter\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
-    }
-
-    protected function getPackageProviders($app)
+    protected function getPackageProviders($app): array
     {
         return [
             LaravelTelemetryReporterServiceProvider::class,
         ];
     }
 
-    public function getEnvironmentSetUp($app)
+    protected function getEnvironmentSetUp($app): void
     {
-        config()->set('database.default', 'testing');
+        // Setup default database to use sqlite :memory:
+        $app['config']->set('database.default', 'testing');
+        $app['config']->set('database.connections.testing', [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+        ]);
 
-        /*
-         foreach (\Illuminate\Support\Facades\File::allFiles(__DIR__ . '/database/migrations') as $migration) {
-            (include $migration->getRealPath())->up();
-         }
-         */
+        // Set up telemetry config
+        $app['config']->set('telemetry-reporter.enabled', true);
+        $app['config']->set('telemetry-reporter.server_url', 'http://example.com/api/telemetry');
+        $app['config']->set('telemetry-reporter.app_host', 'test-host');
+        $app['config']->set('telemetry-reporter.cache_store', 'array');
     }
 }